#45683 Preventing duplicated click for http requets on plan moudle

+++ Teleopti.Ccc.Web/Teleopti.Ccc.Web/WFM/app/resourceplanner/resource_planner_planning_group/planninggroup.createform.controller.js
@@ -9,7 +9,8 @@
 
 	function Controller($state, $timeout, $stateParams, planningGroupService, NoticeService, $translate, debounceService, localeLanguageSortingService, editPlanningGroup) {
 		var vm = this;
-
+		
+		var requestSent = false;
 		vm.searchString = '';
 		vm.selectedResults = [];
 		vm.filterResults = [];
@@ -78,7 +79,7 @@
 				if (node.Id === item.Id) {
 					check = false;
 				};
-			}); 
+			});
 			return check;
 		}
 
@@ -109,14 +110,17 @@
 			if (!isValid()) {
 				NoticeService.warning($translate.instant('CouldNotApply'), 5000, true);
 				return;
+			} else if (!requestSent) {
+				requestSent = true;
+				return planningGroupService.savePlanningGroup({
+					Id: editPlanningGroup ? editPlanningGroup.Id : null,
+					Name: vm.name,
+					Filters: vm.selectedResults
+				}).$promise.then(function () {
+					requestSent = false;
+					$state.go('resourceplanner.newoverview');
+				});
 			}
-			return planningGroupService.savePlanningGroup({
-				Id: editPlanningGroup ? editPlanningGroup.Id : null,
-				Name: vm.name,
-				Filters: vm.selectedResults
-			}).$promise.then(function () {
-				$state.go('resourceplanner.newoverview');
-			});
 		}
 
 		function cancel() {
@@ -125,9 +129,13 @@
 
 		function removePlanningGroup() {
 			if (!editPlanningGroup) return;
-			return planningGroupService.removePlanningGroup({ id: editPlanningGroup.Id }).$promise.then(function () {
-				$state.go('resourceplanner.newoverview');
-			});
+			if (!requestSent) {
+				requestSent = true;
+				return planningGroupService.removePlanningGroup({ id: editPlanningGroup.Id }).$promise.then(function () {
+					requestSent = false;
+					$state.go('resourceplanner.newoverview');
+				});
+			}
 		}
 	}
 })();
